Syncing from Monotone at 20191221-024655. Last commit message was Cleanup of the example

+++ rlwrap-user-shell.sh
@@ -1,9 +1,13 @@
 #! /bin/sh
-touch ~/.user-lisp-shell_history
-history="$(mktemp ~/.user-lisp-shell_history.XXXXXXXX)"
-cat ~/.user-lisp-shell_history > "$history"
-rlwrap -C user-lisp-shell -D 2 -s 1000000  -H "$history" -f . -c -e '' -b "'\"() " -q "\"" "$@"
-cat "$history" >> ~/.user-lisp-shell_history
-cat ~/.user-lisp-shell_history | sort | uniq > ~/.user-lisp-shell_history.new
-mv  ~/.user-lisp-shell_history.new  ~/.user-lisp-shell_history
-rm "$history"
+if which rlwrap &> /dev/null; then
+        touch ~/.user-lisp-shell_history
+        history="$(mktemp ~/.user-lisp-shell_history.XXXXXXXX)"
+        cat ~/.user-lisp-shell_history > "$history"
+        rlwrap -C user-lisp-shell -D 2 -s 1000000  -H "$history" -f . -c -e '' -b "'\"() " -q "\"" "$@"
+        cat "$history" >> ~/.user-lisp-shell_history
+        cat ~/.user-lisp-shell_history | sort | uniq > ~/.user-lisp-shell_history.new
+        mv  ~/.user-lisp-shell_history.new  ~/.user-lisp-shell_history
+        rm "$history"
+else
+        "$@"
+fi
